1. Commit - 12.02.2024
Data Matching and Updating Script (match_and_update.py): This script compares values from column C in 'Sayfa1' of an Excel file with column B in 'Sayfa2'. When a match is found, it updates column B in 'Sayfa1' with the corresponding value from column A in 'Sayfa2'. The script handles direct matches and matches after trimming spaces.

Column Update Script (update_column.py): This script modifies values in column C of 'Sayfa2' in an Excel file. It removes the portion of each string up to the first '-' character, updating the cell with this new, trimmed value.